packages/cli: move buildContract to core
