+++ packages/cli/package.json
@@ -30,7 +30,6 @@
 		"chalk": "^5.4.1",
 		"cors": "^2.8.5",
 		"dotenv": "^16.4.7",
-		"esbuild": "^0.24.2",
 		"ethers": "^6.13.5",
 		"express": "^4.21.2",
 		"http-status-codes": "^2.3.0",
