+++ packages/cli/src/utils.ts
@@ -4,32 +4,13 @@ import process from "node:process"
 
 import { bytesToHex } from "@noble/hashes/utils"
 import { sha256 } from "@noble/hashes/sha256"
-import esbuild from "esbuild"
 import chalk from "chalk"
 import prompts from "prompts"
 
+import { Canvas } from "@canvas-js/core"
+
 export const CONTRACT_FILENAME = "contract.canvas.js"
 export const MANIFEST_FILENAME = "canvas.json"
-
-function buildContract(location: string) {
-	const bundle = esbuild.buildSync({
-		bundle: true,
-		platform: "node",
-		format: "esm",
-		write: false,
-		entryPoints: [location],
-	})
-	if (!bundle.outputFiles || bundle.outputFiles.length === 0) {
-		console.error(chalk.yellow("[canvas] Building .ts contract produced no files"))
-		process.exit(1)
-	} else if (bundle.outputFiles && bundle.outputFiles.length > 1) {
-		console.warn(chalk.yellow("[canvas] Building .ts contract produced more than one file, likely will not run"))
-		return bundle.outputFiles[0].text
-	} else {
-		console.log(chalk.yellow("[canvas] Bundled .ts contract:"), `${bundle.outputFiles[0].contents.byteLength} bytes`)
-		return bundle.outputFiles[0].text
-	}
-}
 
 export function getContractLocation(args: { path: string; topic?: string; init?: string; memory?: boolean }): {
 	topic: string
@@ -53,7 +34,7 @@ export function getContractLocation(args: { path: string; topic?: string; init?:
 			if (args.init.endsWith(".js")) {
 				fs.copyFileSync(args.init, contractPath)
 			} else {
-				fs.writeFileSync(contractPath, buildContract(args.init))
+				fs.writeFileSync(contractPath, Canvas.buildContract(args.init))
 			}
 			console.log(`[canvas] Creating ${manifestPath}`)
 			fs.writeFileSync(manifestPath, JSON.stringify({ version: 1, topic: args.topic }, null, "  "))
@@ -90,7 +71,7 @@ export function getContractLocation(args: { path: string; topic?: string; init?:
 		let contract = fs.readFileSync(location, "utf-8")
 
 		if (location.endsWith(".ts")) {
-			contract = buildContract(location)
+			contract = Canvas.buildContract(location)
 		}
 
 		const topic = args.topic ?? `${bytesToHex(sha256(contract)).slice(0, 16)}.p2p.app`
