+++ packages/core/package.json
@@ -58,6 +58,7 @@
 		"chalk": "^5.4.1",
 		"cors": "^2.8.5",
 		"ethers": "^6.13.5",
+		"esbuild": "^0.24.2",
 		"express": "^4.21.2",
 		"express-ipld": "^0.0.1",
 		"http-status-codes": "^2.3.0",
