+++ packages/core/src/targets/interface.ts
@@ -12,4 +12,6 @@ export interface PlatformTarget {
 	) => Promise<AbstractGossipLog<MessageType>>
 
 	listen: (app: Canvas, port: number, options?: { signal?: AbortSignal }) => Promise<void>
+
+	buildContract: (location: string) => string
 }
