+++ packages/core/src/targets/react-native/index.ts
@@ -12,31 +12,16 @@ const target: PlatformTarget = {
 			const { GossipLog } = await import("@canvas-js/gossiplog/sqlite-expo")
 			return await GossipLog.open({ ...init, clear: location.clear })
 		} else {
-			throw new Error("Unimplemented named sqlite dbs")
+			throw new Error("Unimplemented: named sqlite dbs on react-native")
 		}
 	},
 
 	async listen(app, port, options = {}) {
-		throw new Error("Unimplemented for iOS")
+		throw new Error("Unimplemented: libp2p listen on react-native")
+	},
 
-		// const api = express()
-		// api.use(cors())
-		// api.use("/api", createAPI(app))
-
-		// // TODO: add metrics API
-
-		// const server = http.createServer(api)
-		// const network = new NetworkServer(app.messageLog)
-		// const wss = new WebSocketServer({ server, perMessageDeflate: false })
-		// wss.on("connection", network.handleConnection)
-
-		// const signal = anySignal([app.messageLog.controller.signal, options.signal])
-		// signal.addEventListener("abort", () => {
-		// 	network.close()
-		// 	wss.close(() => server.close())
-		// })
-
-		// await new Promise<void>((resolve) => server.listen(port, resolve))
+	buildContract(location: string): string {
+		throw new Error("Unimplemented: buildContract on react-native")
 	},
 }
